sms-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-tenants==3.5.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.4.0
reportlab==4.2.2
whitenoise==6.12.0
gunicorn==21.2.0
